Add Dockerfile and refactor the codebase
The Dockerfile was added to set up an environment for the application. Refactoring was carried out across the codebase that included renaming data attributes to improve consistency, moving menu_data.ts to a new location within the project, and revising import paths. The application was enhanced with lazy loading of feature modules, updating copy on the home page, and changes to enhance user interaction with todos. The API endpoint in the Todo service was also updated.