Add side menu to header component
A side navigation menu has been added to the header component. This new feature includes a "Menu" button that triggers a dropdown menu with additional links for "Login" and "Register". The relevant adjustments were made in the html template and typescript file for the header component, as well as the `menu_data.ts` and `shared.module.ts` files.